Seperate Expected Results and Actual Results
This commit seperates the expected result type and actual result types. This refactor allows for different types of checks (comparison based on type, based on == operator, based on public properties & fields). This refactor also shifts to using xunit.assert to handle test assertions instead of custom implementations. This means that a. assertion implementations are in a codebase that is actively maintained by .NET and b. Apache 2.0 license is included in Builds as they now contain the xunit dll.